refactor(vitals): move the alert config schema out of the per-customer overlay
`packages/utils/lib/ottehr-config/vitals/index.ts` is replaced wholesale in a
customer build, so the zod schemas, `VitalsDef`, `DefaultVitalsConfig` and
`vitalsConfig` that lived beside the data were duplicated into every customer
config and only reached a customer once their copy was regenerated.

Move all of it to `helpers/vitals/config-schema.ts`, which is not overlaid, and
leave `ottehr-config/vitals/index.ts` holding `VitalsConfigData` alone. Repoint
the ten importers. Deliberately not re-exported from any barrel: `main.ts`
re-exports both `./helpers` and `./ottehr-config`, so a customer still shipping
the old file would otherwise export `VitalsDef` twice.

Compatible in both directions, verified against simulated customer builds: a
config that still carries thresholds is read as before, and a data-only one
falls back to the shipped defaults with an error logged.
