+++ apps/ehr/src/features/visits/shared/components/vitals/weights/VitalsWeightsCard.tsx
@@ -5,8 +5,8 @@ import { AccordionCard } from 'src/components/AccordionCard';
 import { DoubleColumnContainer } from 'src/components/DoubleColumnContainer';
 import { RoundedButton } from 'src/components/RoundedButton';
 import { dataTestIds } from 'src/constants/data-test-ids';
+import { vitalsConfig } from 'utils/lib/helpers/vitals/config-schema';
 import { formatWeight, formatWeightKg, formatWeightLbs } from 'utils/lib/helpers/vitals/vitals-weight.helper';
-import { vitalsConfig } from 'utils/lib/ottehr-config/vitals';
 import { VitalFieldNames } from 'utils/lib/types/api/chart-data/chart-data.constants';
 import { VitalsWeightObservationDTO, VitalsWeightOption } from 'utils/lib/types/api/chart-data/chart-data.types';
 import { useGetAppointmentAccessibility } from '../../../hooks/useGetAppointmentAccessibility';
