+++ packages/utils/lib/utils/vitals-alert-config.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { VitalsAge } from '../config-helpers/vitals';
-import { VitalsDef, VitalsSchema } from '../ottehr-config/vitals';
+import { VitalsDef, VitalsSchema } from '../helpers/vitals/config-schema';
 import { VitalAlertCriticality } from '../types/api/chart-data/chart-data.constants';
 import {
   VITAL_ALERT_TYPES,
